feat: move profile page to be part of auth
