--- app/routes/me/page.tsx
@@ -1,9 +1,0 @@
-import { useLoaderData } from "react-router-dom";
-
-function Page() {
-  const profile = useLoaderData();
-  console.log(profile);
-  return <div>Profile</div>;
-}
-
-export { Page };
