+++ app/routes/root.tsx
@@ -6,7 +6,9 @@ function Root() {
     <main className="h-screen w-screen flex flex-col">
       <NavigationMenu />
       <div className="grow shrink-0">
-        <Outlet />
+        <div className="bg-gradient-to-t from-indigo-500 via-purple-500 to-pink-500 h-full w-full">
+          <Outlet />
+        </div>
       </div>
     </main>
   );
